extract function and add test
